# src/concord/server/models/channel_related_response.py
# coding: utf-8

"""
    Concord API

    API for Concord, an AI-powered semantic extraction and recommendation platform for networked communities.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501

from __future__ import annotations

import json
import pprint
import re  # noqa: F401
from typing import Any, ClassVar, Dict, List, Optional

from pydantic import BaseModel, StrictStr

from src.concord.server.models.channel_related_response_related_channels_inner import \
    ChannelRelatedResponseRelatedChannelsInner

try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class ChannelRelatedResponse(BaseModel):
    """
    ChannelRelatedResponse
    """

    # noqa: E501
    platform_id: Optional[StrictStr] = None
    channel_id: Optional[StrictStr] = None
    related_channels: Optional[
        List[ChannelRelatedResponseRelatedChannelsInner]] = None
    __properties: ClassVar[List[str]] = [
        "platform_id", "channel_id", "related_channels"
    ]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ChannelRelatedResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={},
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in related_channels (list)
        _items = []
        if self.related_channels:
            for _item in self.related_channels:
                if _item:
                    _items.append(_item.to_dict())
            _dict['related_channels'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ChannelRelatedResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "platform_id":
                obj.get("platform_id"),
            "channel_id":
                obj.get("channel_id"),
            "related_channels": [
                ChannelRelatedResponseRelatedChannelsInner.from_dict(_item)
                for _item in obj.get("related_channels")
            ] if obj.get("related_channels") is not None else None
        })
        return _obj