# src/concord/server/models/server_register_request.py
# coding: utf-8

"""
    Concord API

    API for Concord, an AI-powered semantic extraction and recommendation platform for networked communities.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501

from __future__ import annotations

import json
import pprint
import re  # noqa: F401
from typing import Any, ClassVar, Dict, List, Optional

from pydantic import BaseModel, Field, StrictStr, field_validator

try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class ServerRegisterRequest(BaseModel):
    """
    ServerRegisterRequest
    """

    # noqa: E501
    platform: Optional[StrictStr] = Field(default=None,
                                          description="Name of the platform.")
    name: Optional[StrictStr] = Field(default=None,
                                      description="Name of the server/group.")
    auth_token: Optional[StrictStr] = Field(
        default=None, description="Authentication token.")
    description: Optional[StrictStr] = Field(
        default=None, description="Optional description of the server/group.")
    contact_email: Optional[StrictStr] = Field(
        default=None, description="Optional contact email.")
    webhook_url: Optional[StrictStr] = Field(
        default=None, description="Optional URL for webhook updates.")
    __properties: ClassVar[List[str]] = [
        "platform", "name", "auth_token", "description", "contact_email",
        "webhook_url"
    ]

    @field_validator('platform')
    def platform_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value

        if value not in ('matrix', 'slack', 'telegram', 'discord'):
            raise ValueError(
                "must be one of enum values ('matrix', 'slack', 'telegram', 'discord')"
            )
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ServerRegisterRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={},
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ServerRegisterRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "platform": obj.get("platform"),
            "name": obj.get("name"),
            "auth_token": obj.get("auth_token"),
            "description": obj.get("description"),
            "contact_email": obj.get("contact_email"),
            "webhook_url": obj.get("webhook_url")
        })
        return _obj